+++ packages/@expo/cli/src/start/server/metro/metroErrorInterface.ts
@@ -278,6 +278,7 @@ export async function getErrorOverlayHtmlAsync({
       mode: 'development',
       platform: 'web',
       minify: false,
+      optimize: false,
       baseUrl: '',
       routerRoot,
       isExporting: false,